added cast to prevent compile error in gcc
git-svn-id: svn://svn.pgadmin.org/trunk@1326 a7884b65-44f6-0310-8a51-81a127f17b15

+++ src/db/pgSet.cpp
@@ -120,7 +120,7 @@ long pgSet::GetLong(int col) const
 
 long pgSet::GetLong(const wxString &col)
 {
-    return GetLong(ColNumber(col));
+    return GetLong((long)ColNumber(col));
 }
